Fix 64-bit portability warnings and type conversions
Address critical warnings for 32→64 bit architecture port:

- Fix format specifiers: use %ld for long, %lx for Window/Drawable
- Add explicit casts for safe type conversions:
  * size_t → int where values are bounded (parser buffers, counts)
  * int → short for font metrics (maxFontHeight, maxFontDescent, breakc)
  * int → char for font/tag IDs in TFChar structures
  * long → unsigned long for X11 types (Window, Drawable)
- Fix malloc() calls to use size_t for allocation sizes
- Fix pointer type mismatches in biop.c array initialization
  (use & to get pointer to entire array vs pointer to first element)
- Add missing X11/Intrinsic.h include for XtDispatchEvent

Files changed:
- src/viola/ast.c: Fix strlen result handling in lexer
- src/viola/biop.c: Fix ConvIdx array initialization with proper pointers
- src/viola/cexec.c: Fix format strings, malloc sizes, pointer casts
- src/viola/html.c: Fix TFLineNode allocations and type conversions
- src/viola/tfed.c: Fix font metrics assignments and debug output
- src/vw/mail.c: Fix string length calculation
- src/vw/history.c: Fix XmStringTable allocations
- src/vw/hotlist.c: Fix XmCreateScrolledList argument types

All changes preserve functionality while ensuring correct behavior
on 64-bit platforms.

+++ src/vw/mail.c
@@ -314,7 +314,7 @@ XtPointer clientData, callData;
     address = XmTextFieldGetString(ms->addressText);
     subject = XmTextFieldGetString(ms->subjectText);
     letter = XmTextGetString(ms->letterText);
-    length = strlen(mailCommand) + strlen(address) + strlen(subject) + 20;
+    length = (int)(strlen(mailCommand) + strlen(address) + strlen(subject) + 20);
     sprintf(fullCommand, "%s %s %s", mailCommand, subject, address);
     sprintf(dlogMessage, "Sending your letter using the following command:\n\n%s\n\nIs this OK?",
             fullCommand);
